Add files via upload
This project is a pet store website, which admins can add, update, and delete products from the website, and users can add pet products to the cart and buy buy pets or pets supplies. The actual payment is not included in this website since this is a university project which was conducted with a team.